# Speed up build times and encapsulate github auth