--- a/pkg/retry/retry.go
+++ b/pkg/retry/retry.go
@@ -23,7 +23,7 @@ import (
 	"time"
 
 	"github.com/abcxyz/pkg/logging"
-	"github.com/google/go-github/v52/github"
+	"github.com/google/go-github/v56/github"
 	"github.com/sethvargo/go-gcslock"
 )