--- a/pkg/retry/server.go
+++ b/pkg/retry/server.go
@@ -24,7 +24,7 @@ import (
 
 	"github.com/abcxyz/pkg/healthcheck"
 	"github.com/abcxyz/pkg/logging"
-	"github.com/google/go-github/v52/github"
+	"github.com/google/go-github/v56/github"
 	"github.com/sethvargo/go-gcslock"
 	"google.golang.org/api/option"
 )