@pdewilde 2023-11-13
Just a general question about why the init() causes such a build time difference:

Is it because beam already has a lot of `init()` and its non linear to add another in the ordering decision?