@sethvargo 2023-11-14
> Just a general question about why the init() causes such a build time difference:
> 
> Is it because beam already has a lot of `init()` and its non linear to add another in the ordering decision?

I think there are a few things going against us:

- Go has to order the package imports and _then_ run all inits, because it needs to run init functions in their import order. Each init function is _blocking_, so a single init function can prevent the entire function from booting.

- init functions themselves tend to allocate, often a lot. They establish database connections, make upstream calls, and _cannot_ be context-aware.

-